# sink: reduce lock competition during flushBackendSink