--- a/cdc/sink/manager.go
+++ b/cdc/sink/manager.go
@@ -99,7 +99,7 @@ func (m *Manager) flushBackendSink(ctx context.Context) (model.Ts, error) {
 	// which will cause a lot of lock contention and blocking in high concurrency cases.
 	// So here we use flushing as a lightweight lock to improve the lock competition problem.
 	if !atomic.CompareAndSwapInt64(&m.flushing, 0, 1) {
-		return atomic.LoadUint64(&m.checkpointTs), nil
+		return m.getCheckpointTs(), nil
 	}
 	m.flushMu.Lock()
 	defer func() {