--- a/cdc/sink/manager_test.go
+++ b/cdc/sink/manager_test.go
@@ -249,64 +249,64 @@ func (s *managerSuite) TestManagerDestroyTableSink(c *check.C) {
 }
 
 func BenchmarkManagerRandomFlushing(b *testing.B) {
-	for n := 0; n < b.N; n++ {
-		ctx, cancel := context.WithCancel(context.Background())
-		errCh := make(chan error, 16)
-		manager := NewManager(ctx, &checkSink{}, errCh, 0)
+	ctx, cancel := context.WithCancel(context.Background())
+	errCh := make(chan error, 16)
+	manager := NewManager(ctx, &checkSink{}, errCh, 0)
 
-		goroutineNum := 1000
-		rowNum := 100
-		var wg sync.WaitGroup
-		tableSinks := make([]Sink, goroutineNum)
-		for i := 0; i < goroutineNum; i++ {
-			i := i
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
-				tableSinks[i] = manager.CreateTableSink(model.TableID(i), 0)
-			}()
-		}
-		wg.Wait()
-		for i := 0; i < goroutineNum; i++ {
-			i := i
-			tableSink := tableSinks[i]
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
-				ctx := context.Background()
-				var lastResolvedTs uint64
-				for j := 1; j < rowNum; j++ {
-					if rand.Intn(10) == 0 {
-						resolvedTs := lastResolvedTs + uint64(rand.Intn(j-int(lastResolvedTs)))
-						_, err := tableSink.FlushRowChangedEvents(ctx, resolvedTs)
-						if err != nil {
-							b.Error(err)
-						}
-						lastResolvedTs = resolvedTs
-					} else {
-						err := tableSink.EmitRowChangedEvents(ctx, &model.RowChangedEvent{
-							Table:    &model.TableName{TableID: int64(i)},
-							CommitTs: uint64(j),
-						})
-						if err != nil {
-							b.Error(err)
-						}
+	goroutineNum := 1000
+	rowNum := 100
+	var wg sync.WaitGroup
+	tableSinks := make([]Sink, goroutineNum)
+	for i := 0; i < goroutineNum; i++ {
+		i := i
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			tableSinks[i] = manager.CreateTableSink(model.TableID(i), 0)
+		}()
+	}
+	wg.Wait()
+	for i := 0; i < goroutineNum; i++ {
+		i := i
+		tableSink := tableSinks[i]
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			ctx := context.Background()
+			var lastResolvedTs uint64
+			for j := 1; j < rowNum; j++ {
+				if rand.Intn(10) == 0 {
+					resolvedTs := lastResolvedTs + uint64(rand.Intn(j-int(lastResolvedTs)))
+					_, err := tableSink.FlushRowChangedEvents(ctx, resolvedTs)
+					if err != nil {
+						b.Error(err)
+					}
+					lastResolvedTs = resolvedTs
+				} else {
+					err := tableSink.EmitRowChangedEvents(ctx, &model.RowChangedEvent{
+						Table:    &model.TableName{TableID: int64(i)},
+						CommitTs: uint64(j),
+					})
+					if err != nil {
+						b.Error(err)
 					}
 				}
+			}
+			for n := 0; n < b.N; n++ {
 				_, err := tableSink.FlushRowChangedEvents(ctx, uint64(rowNum))
 				if err != nil {
 					b.Error(err)
 				}
-			}()
-		}
-		wg.Wait()
-		cancel()
-		_ = manager.Close(ctx)
-		close(errCh)
-		for err := range errCh {
-			if err != nil {
-				b.Error(err)
 			}
+		}()
+	}
+	wg.Wait()
+	cancel()
+	_ = manager.Close(ctx)
+	close(errCh)
+	for err := range errCh {
+		if err != nil {
+			b.Error(err)
 		}
 	}
 }