@hi-rustin 2021-09-06
Benchmark result: 

Before:

```sh
➜  sink git:(rustin-patch-fix-sink) ✗ go test  -bench='^\QBenchmarkManagerRandomFlushing\E$' -test.run=thisexpressionwontmatchanytest -count=10 .
goos: darwin
goarch: amd64
pkg: github.com/pingcap/ticdc/cdc/sink
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkManagerRandomFlushing-16              1        4466272548 ns/op
BenchmarkManagerRandomFlushing-16              1        4692509999 ns/op
BenchmarkManagerRandomFlushing-16              1        4566925133 ns/op
BenchmarkManagerRandomFlushing-16              1        4553761846 ns/op
BenchmarkManagerRandomFlushing-16              1        4628824677 ns/op
BenchmarkManagerRandomFlushing-16              1        4725037126 ns/op
BenchmarkManagerRandomFlushing-16              1        4786872104 ns/op
BenchmarkManagerRandomFlushing-16              1        4513431907 ns/op
BenchmarkManagerRandomFlushing-16              1        4583744744 ns/op
BenchmarkManagerRandomFlushing-16              1        4616345649 ns/op
PASS
ok      github.com/pingcap/ticdc/cdc/sink       47.078s
```

After:
```sh
➜  sink git:(rustin-patch-fix-sink) ✗ go test  -bench='^\QBenchmarkManagerRandomFlushing\E$' -test.run=thisexpressionwontmatchanytest -count=10 .
goos: darwin
goarch: amd64
pkg: github.com/pingcap/ticdc/cdc/sink
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkManagerRandomFlushing-16             20          55283572 ns/op
BenchmarkManagerRandomFlushing-16             24          53067853 ns/op
BenchmarkManagerRandomFlushing-16             21          53231288 ns/op
BenchmarkManagerRandomFlushing-16             20          51749307 ns/op
BenchmarkManagerRandomFlushing-16             24          52132987 ns/op
BenchmarkManagerRandomFlushing-16             25          52443937 ns/op
BenchmarkManagerRandomFlushing-16             22          52733766 ns/op
BenchmarkManagerRandomFlushing-16             19          53113452 ns/op
BenchmarkManagerRandomFlushing-16             21          53181457 ns/op
BenchmarkManagerRandomFlushing-16             24          53821486 ns/op
PASS
ok      github.com/pingcap/ticdc/cdc/sink       13.207s
```